refactor(workspace): split message queue controller into admission and drain
The architecture gate failed because the React controller exceeded 700 lines.
Keep the public hook and provider as the facade, and move admission policy,
drain/send-now scheduling, active-session projection, and live-region copy
into separate modules around the existing in-memory owner.
