+++ scripts/ci/module-impact.json
@@ -119,6 +119,10 @@
         "src/renderer/src/pages/workspace/workspace-conversation-controller.ts",
         "src/renderer/src/pages/workspace/workspace-message-queue-controller.ts",
         "src/renderer/src/pages/workspace/workspace-message-queue-owner.ts",
+        "src/renderer/src/pages/workspace/workspace-message-queue-admission.ts",
+        "src/renderer/src/pages/workspace/workspace-message-queue-drain.ts",
+        "src/renderer/src/pages/workspace/workspace-message-queue-projection.ts",
+        "src/renderer/src/pages/workspace/workspace-message-queue-announcement.ts",
         "src/renderer/src/pages/workspace/use-workspace-branch-switch-guard.ts",
         "src/renderer/src/pages/workspace/use-side-chat-controller.ts",
         "src/renderer/src/pages/workspace/workspace-session-controller.ts",
